completed the method changes
