--- index.js
@@ -1,3 +1,0 @@
-const {deterministicPartitionKey} = require("./dpk");
-
-console.log(deterministicPartitionKey());
