+++ package.json
@@ -2,14 +2,15 @@
   "name": "cbh-take-home-project",
   "version": "1.0.0",
   "description": "",
-  "main": " ",
+  "main": "src/index.js",
   "scripts": {
     "test": "jest",
-    "start": "node index.js"
+    "start": "node src"
   },
-  "author": "",
+  "author": "Ritesh Firodiya",
   "license": "ISC",
   "devDependencies": {
-    "jest": "^28.1.0"
+    "jest": "^28.1.0",
+    "typescript": "^4.9.3"
   }
 }
