+++ src/index.js
@@ -1,0 +1,3 @@
+const { deterministicPartitionKey } = require("./dpk");
+
+console.log(deterministicPartitionKey());
